+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 > This page is part of the [App Framework Documentation](DOCUMENTATION.md)
 
+## Version 1.20.0
+
+Released on 2018-01-10
+
+### New features
+
+- [x] [#763 - Allow debug mode for App Framework](https://github.com/scriptPilot/app-framework/issues/763)
+
+### Bug-fixes
+
+- [x] [#743 - Update framework7-vue to v0.9.4 to fix popup issue](https://github.com/scriptPilot/app-framework/issues/743)
+- [x] [#762 - Error "Cannot find Framework7-Vue folder"](https://github.com/scriptPilot/app-framework/issues/762)
+
 ## Version 1.19.2
 
 Release 2018-01-05